Extract search index and query evaluation
