+++ crates/truthdb-core/src/engine/search/mod.rs
@@ -1,1 +1,7 @@
 //! Native document-search implementation owned by [`super::Engine`].
+
+mod index;
+
+pub(super) use index::{
+    Document, EngineState, FieldType, IndexState, SearchQuery, value_type_name,
+};
