+++ crates/truthdb-core/src/engine/tests.rs
@@ -1,4 +1,3 @@
-
 use std::path::{Path, PathBuf};
 
 use super::*;
